SFTP.php : prevent endless loop when upload completed with non-zero $start or $local_start
in SFTP::put(), if either of the $start or $local_start parameters is non-zero, and $data is not a callable, then the while {} loop which sends data to the remote server will never terminate, even when the end-of-file on the file being read has been reached. This is because the conidition $sent < $size is never satisfied - because it does not take into account the non-zero start. As a result, when the end of the file has been reached, an endless loop of inactivity will result - the method will never return.