don't run unit tests on PHP 5.6
phpseclib 3.0 still nominally supports PHP 5.6 but i'm tired of seeing
unit tests fail because of issues i can't reproduce on my own local
PHP 5.6 install.

this isn't without precident.phpseclib 1.0 nominally supports PHP 4.4,
as well, but it's not unit tested on PHP 4.4. for that one, if someone
has an issue with PHP 4.4 they're free to report it on github. PHP 5.6
will henceforth work that way for phpseclib 3.0.