fix PHP 8.4 deprecations
phpseclib 3.0 works on PHP 5.6+. changing this for the 3.0 branch
would be a BC break so that's not happening.

the problem is that PHP 8.4 deprecated implicitly nullable parameter
types [1] in favor of the ?T syntax, which wasn't introduced until
PHP 7.1. if phpseclib 3.0 ups the minimimally required PHP version
to PHP 5.6 it's breaking BC so we'll just remove the type hinting
to make it a non issue. the master branch (which will eventually
someday become phpseclib 4.0) will still support type hinting

[1] https://wiki.php.net/rfc/deprecate-implicitly-nullable-types