ASN1: copy DN_* constants from DN trait
in the documentation i'm writing distinguished names have their own
page. like there's enough to write about them such that duplicating that
info on each of the pages for X509s, CSRs and CRLs is most unappealing.
in light of that having a class agnostic way to reference the DN_*
constants is appealing